`reference` folder contains files related to Result import, detailed resutl and gradehistory. student feedback analysis and project fair related components, views and apis.
i want you to use this components, and then adapt it to suite next js app.

feedback analysis and result import are well tested in MERN stack so i dont want to loose any details from this implementation just styyling as per our current next app is suggested and you do.

for not missing any implementation parts, i want you to copy those files as it is to the respective suitable folders and then try to adapt it minimally so that it worked in our next js app.

as they are already from React and express they will not required many changes